[Rename] 페이지별 폴더 정리
