+++ src/App.js
@@ -4,12 +4,12 @@ import RootLayout from '@pages/RootLayout';
 import ErrorPage from '@pages/Error';
 
 import HomeRootLayout from '@pages/HomeRootLayout';
-import HomePage from '@pages/Home';
-import PostDetailPage from '@pages/PostDetail';
-import NotificationPage from '@pages/Notification';
+import HomePage from '@pages/home/Home';
+import PostDetailPage from '@pages/home/PostDetail';
+import NotificationPage from '@pages/home/Notification';
 
 import AuthenticationPage from '@pages/auth/Authentication';
-import { loader as logoutLoader } from '@pages/Logout';
+import { loader as logoutLoader } from '@pages/auth/Logout';
 import SignInPage, { action as signInAction } from '@pages/auth/SignIn';
 import SignUpPage, { action as signUpAction } from '@pages/auth/SignUp';
 import AccountPage from '@pages/auth/Account';
@@ -37,7 +37,7 @@ const router = createBrowserRouter([
             element: <HomePage />,
           },
           {
-            path: '/post/:postId',
+            path: ':postId',
             element: <PostDetailPage />,
           },
           {
